pyproject.toml
[project]
# deps managed by uv
name = "octopus"
version = "0.1.0"
description = "CICD infra for SDLC"
requires-python = ">=3.10, <4.0"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
    "requests>=2.31.0",
    "docker>=7.1.0",
    "jinja2>=3.1.6",
    "python-gitlab>=5.6.0",
    "PyGithub>=2.6.1",
    "python-jenkins>=1.8.2",
    "requests-toolbelt>=1.0.0",
    "poetry>=2.1.3",
    "ruff>=0.11.0",
    "pre-commit>=4.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.1.0",
    "cryptography>=45.0.3",
    "virtualenv>=20.31.2",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
# deps managed by uv
dev = [
    "ruff>=0.11.0",
    "pre-commit>=4.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.1.0",
    "cryptography>=45.0.3",
    "virtualenv>=20.31.2"
]

[tool.poetry]
name = "octopus"
version = "0.1.0"
description = "CICD infra for SDLC"
authors = ["linview@gmail.com"]
packages = []

[tool.poetry.dependencies]
# deps managed by poetry
python = ">=3.10, <4.0"
pyyaml = "^6.0"
click = "^8.0"
pydantic = "^2.0"
requests = "^2.31.0"
docker = "^7.1.0"
jinja2 = "^3.1.6"
python-gitlab = "^5.6.0"
PyGithub = "^2.6.1"
python-jenkins = "^1.8.2"
requests-toolbelt = "^1.0.0"

[tool.poetry.group.dev.dependencies]
# deps managed by poetry
ruff = "^0.11.0"
pre-commit = "^4.2.0"
pytest = "^8.0.0"
pytest-cov = "^6.1.0"
cryptography = "^45.0.3"
virtualenv = "^20.31.2"

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "N", "W", "B", "UP", "C"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.uv]
index-url = "https://pypi.org/simple"
